The program have now a simple GUI.
It can update the local DB (which is basically a text file). It can search words in the DB too. Only problem is that for now the Text Area on the GUI don't update as expected when new text is entered.
Changed much in FichierW and FichierR too (in methods and globally their behavior)